requirements.txt
Flask
PyMySQL
flask-cors
Flask-SQLAlchemy
gunicorn
